requirements.txt
dynaconf==3.1.11
mysql-connector-python==8.0.31
protobuf==3.20.1
